### 2020-05-26 #2255 @danielskatz
👋 @edwardoughton - Please suggest some potential reviewers (without tagging them with an @).

[This list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).

Or if you have other suggestions for potential reviewers, then please mention them here.
