### 2020-05-26 #2255 @edwardoughton
Thanks for the guidance Daniel.

Potential reviewer - alirazaanis - creator of the community cellular planner repo.
