### 2020-05-26 #2255 @edwardoughton
Potential reviewer: gorlapraveen (previous JOSS reviewer - knows radio/electrical engineering etc. - uses python)
