### 2020-05-26 #2255 @edwardoughton
Potetial reviewer: garrettj403 (previous JOSS reviewer - knows radio/electrical engineering etc. - uses python)
