### 2020-05-27 #2255 @edwardoughton
Potential reviewer: peterl3233 (RF engineer who also codes)
