### 2020-05-27 #2255 @gorlapraveen
> wave @gorlapraveen - Would you be willing to review this submission for JOSS?

@danielskatz I wish to review!
