### 2020-05-28 #2266 @whedon
[REVIEW]: itmlogic: Longley Rice Irregular Terrain Model
**Submitting author:** @edwardoughton (<a href="http://orcid.org/0000-0002-2766-008X">Edward J. Oughton</a>)
**Repository:** <a href="https://github.com/edwardoughton/itmlogic" target ="_blank">https://github.com/edwardoughton/itmlogic</a>
**Version:** v1.1
