### 2020-05-28 #2266 @danielskatz
👋 @gorlapraveen & @garrettj403 - thanks for agreeing to review this submission. Please be sure to read the comments above, and let me know if you have any questions.  Basically, your job is to check the article proof and repository and check items off your checklist above.

If you see small problems that need to be discussed, feel free to discuss them here.  But if you can, create a new issue in the <a href="https://github.com/edwardoughton/itmlogic" target="_blank">target repository</a> and linking to this review thread in that issue to create corresponding breadcrumb trail here.

I look forward to seeing how this review goes
