### 2020-05-31 edwardoughton/itmlogic#34 @gorlapraveen
Packages Needed
`Traceback (most recent call last):
  File "scripts/p2p.py", line 28, in <module>
    from terrain_module import terrain_p2p
  File "/itmlogic/scripts/terrain_module.py", line 21, in <module>
    from rasterstats import gen_zonal_stats, point_query
ModuleNotFoundError: No module named 'rasterstats'
`

To suppress this Error need to add `rasterstats` and `simplejson` packages installation through conda/pip . And these are not optional as I observed to test p2p mode script.
