+++ .travis.yml
@@ -23,8 +23,6 @@ notifications:
 deploy:
   provider: pypi
   user: edwardoughton
-  password:
-    secure: r71kI38ie486F2ymbaGSUnXbegiZ3Sku7jPjVmmNFxNx+/OyymdApzuECqLWgWQB5S1IdH97C/zg3v+2wbxxgyI7gnFBhTuE4Frn4RwuMWmnRe/6rlUDeV5On1N5DepNmdleLA9aoJPXKjYuaULCHdsql+mCZsRL0eJ8XoHjutLXZdPxK+sS9xvzexTsEvHk6tiu7d7Yg+foH4Xp6dgpGnCeIS9cJ0VrpfY/oed6gkrm0V/fzKLbX9emndPSEnZvlnbYxYJU2J8DzqnztU78wrquQWadkZHK5XLRxGaKexOeUTo+oRCTCzf0fAQoxbM7pWa0p6PRkmT/IHeaSqeAxr1c0gE8JOdLCMwM32Fyi77ko3hPSkqYzbiOP/zNxDUVG9ZMdb5mvEIs/uvCPG3p0YYm4SXEGnbPIzHaKPef0vD45abS8b0wDH9GsLqrz44iKzpripZ2FGuVjE5vEr3yHEMjLSZiISkdhuNSMf1Nz2JwYpocRVEdmVVN8EpXmLArBHq8NVzb0IVDqFNRkiczPyOref0H4nEG0zP9/ye8CtIPfWYawHLDFr7vJasT63y3pASVOAu31eVdZ0SPBUAqJDzR9e+JhQmILBdDiGSYX8eB8gj7SFgQEDrXo9FFKQqE2V4TRTbM5W4vSfUgLO3CDvvt803uTiIb9L642zsmzlY=
   on:
     tags: true
     branch: master
