### 2020-06-12 #2266 @danielskatz
Thanks @garrettj403  - if you and @gorlapraveen can add a link to this issue (https://github.com/openjournals/joss-reviews/issues/2266) in any issues you open in the source repo, the will be flagged here and anyone looking at this issue will be able to tell if they are open or closed as well.
