### 2020-06-12 edwardoughton/itmlogic#34 @danielskatz
found in JOSS review https://github.com/openjournals/joss-reviews/issues/2266
