### 2020-06-19 edwardoughton/itmlogic#34 @edwardoughton
I have now added both packages to the conda install / pip reqs, and then tested with a clean install. This should now be solved.
