### 2020-06-19 #2266 @edwardoughton
> Hello @edwardoughton,
>
> I have a couple comments on your paper:
>
> * You wrote "... access to disparate user base ...". Do you mean "dispersed user base"?
> * I think that "Longley-Rice" should be hyphenated.
> * You should check the capitalization of the titles in your references. E.g., "The its irregular terrain model..." should be changed to "The ITS Irregular Terrain Model...".
> * The final reference is wonky. You have the Institute for Telecommunication Sciences listed as the author.
>
> Overall, the paper was very well-written. The software was also easy to install on my computer and I hope to finish reviewing the code in the next couple days. Note: I had the same issue as @gorlapraveen that he raised in [issue #34](https://github.com/edwardoughton/itmlogic/issues/34) (missing `rasterstats` package).

@garrettj403 I've now made most of those changes. I'm not sure the reference is wonky for the last entry in the bib as this is a citation of the website that releases the code. I think ITS seems like the right author, as it's their website, but I'm happy to be convinced otherwise.
