+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{Oughton:2019a,
-    title = {Assessing the capacity, coverage and cost of 5G infrastructure strategies: Analysis of the Netherlands},
+    title = {Assessing the capacity, coverage and cost of {5G} infrastructure strategies: Analysis of the {N}etherlands},
     journal = {Telematics and Informatics},
     volume = {37},
     pages = {50 - 69},
@@ -10,7 +10,7 @@
     author = {{Oughton}, E.~J. and {Frias}, Z. and {van der Gaast}, S. and {van der Berg}, R.},
 }
 @article{Oughton:2019b,
-    title = {An open-source techno-economic assessment framework for 5G deployment},
+    title = {An open-source techno-economic assessment framework for {5G} deployment},
     journal = {IEEE Access},
     pages = {1 - 1},
     year = {2019},
@@ -20,7 +20,7 @@
     author = {{Oughton}, E.~J. and {Katsaros}, K. and {Entezami}, F. and {Kaleshi}, D. and {Crowcroft}, J.},
 }
 @article{Hufford:1982,
-    title = {A guide to the use of the ITS Irregular Terrain Model in the area prediction mode},
+    title = {A guide to the use of the {ITS} Irregular Terrain Model in the area prediction mode},
     journal = {NTIA Report},
     pages = {1 - 67},
     year = {1982},
@@ -28,7 +28,7 @@
     author = {{Hufford}, G.~A. and {Longley}, A.~G. and {Kissick}, W.~A.},
 }
 @article{Hufford:1995,
-    title = {The ITS Irregular Terrain Model, version 1.2.2, the Algorithm},
+    title = {The {ITS} Irregular Terrain Model, version 1.2.2, the Algorithm},
     journal = {NTIA Report},
     pages = {1 - 16},
     year = {1995},
@@ -44,8 +44,8 @@
     author = {{Longley}, A.~G.},
 }
 @misc{ITS:2007,
-    author = {Institute for Telecommunication Sciences},
-    title = {Irregular Terrain Model (ITM) (Longley-Rice) (20 MHz – 20 GHz)},
+    author = {{Institute for Telecommunication Sciences}},
+    title = {Irregular Terrain Model ({ITM}) ({Longley-Rice}) (20 {MHz} – 20 {GHz})},
     year = {2007},
     url = {https://www.its.bldrdoc.gov/resources/radio-propagation-software/itm/itm.aspx},
     note = {[Online; accessed 16-March-2020]},
