### 2020-06-19 #2266 @danielskatz
the bib problem is in formatting - please merge https://github.com/edwardoughton/itmlogic/pull/39 to fix it (and some other bib problems)
