### 2020-06-19 #2266 @danielskatz
then you can put in a new comment here with ```@whedon generate pdf``` to regenerate the pdf.
