### 2020-06-19 #2266 @edwardoughton
@gorlapraveen @garrettj403 @danielskatz I think that fixed the bibliography ordering plus capitalizations. Please let me know if there is anything else.
