### 2020-06-19 #2266 @danielskatz
I'm not sure why you made changes manually rather than just merging the PR.  You missed the {}s around MHz and GHz.  Otherwise, it looks ok to me.
