+++ paper/paper.bib
@@ -45,7 +45,7 @@
 }
 @misc{ITS:2007,
     author = {{Institute for Telecommunication Sciences}},
-    title = {{I}rregular {T}errain {M}odel ({ITM}) ({L}ongley-{R}ice) (20 MHz – 20 GHz)},
+    title = {{I}rregular {T}errain {M}odel ({ITM}) ({L}ongley-{R}ice) (20 {MH}z – 20 {GH}z)},
     year = {2007},
     url = {https://www.its.bldrdoc.gov/resources/radio-propagation-software/itm/itm.aspx},
     note = {[Online; accessed 16-March-2020]},
