### 2020-06-19 #2266 @edwardoughton
> I'm not sure why you made changes manually rather than just merging the PR. You missed the {}s around MHz and GHz. Otherwise, it looks ok to me.

Apologies Daniel, I missed the PR! Thanks for that. I hope this is resolved now.
