### 2020-06-22 edwardoughton/itmlogic#43 @garrettj403
Additional comments in example scripts
I'm having a hard time following the examples in the ``scripts/`` directory. Could you please add a few more inline comments? For example, ``scripts/area.py`` doesn't have any comments past line 107.

You may also want to consider adding a quick script to plot the results.

Note that there's a typo in README on line 71: ``scripts/uarea.py`` should be ``scripts/area.py``.
