### 2020-06-22 edwardoughton/itmlogic#44 @garrettj403
Docstrings for tests
It's very difficult to follow what the tests are testing.

For example, in ``test_lrprop.py``, where does ``setup_expected_prop_to_test_lrprop`` come from? Were these results generated by the previous FORTRAN/C++/MATLAB code?

It would be helpful to have short docstrings broadly describing each test.
