### 2020-06-24 edwardoughton/itmlogic#43 @edwardoughton
@garrettj403 thanks for the detailed review, I'm going to turn this around by the weekend.
