### 2020-06-26 edwardoughton/itmlogic#45 @tomalrussell
Consider approachability and legibility of code
Capturing comments by @garrettj403 in the course of JOSS review.

Essential TODOs:
- [ ] [#43](https://github.com/edwardoughton/itmlogic/issues/43) comments in scripts
- [ ] [#44](https://github.com/edwardoughton/itmlogic/issues/44) docstrings in tests
- [x] fix [failing build](https://travis-ci.org/github/edwardoughton/itmlogic/builds/700223839)

Broader recommendation is to refactor and rename away from terse original variable names, and combine the subroutines into related modules and subpackages, with the goal of making the whole package a lot easier to read through and understand.

> I like that you provided two examples in the ``scripts/`` directory, but I think that they need a bit more explanation. For example, ``scripts/area.py`` doesn't have any inline comments past line 107, which makes it difficult to follow along. I raised this issue in [#43](https://github.com/edwardoughton/itmlogic/issues/43).
>
> The testing looks good (I think), but again it's very difficult to follow along without any comments. For example, I'm looking at the test for ``lrprop`` (``test_lrprop.py``). Where does ``setup_expected_prop_to_test_lrprop`` come from? Were these results generated by the previous FORTRAN/C++/MATLAB code? I think that you should have a short docstring in each file to describe the general idea behind the test. I raised this issue in [#44](https://github.com/edwardoughton/itmlogic/issues/44).
>
> Outside the scope of the JOSS review, I have a few general comments about the package:
>   - The variable names are very short, usually only 2 or 3 letters. Python is somewhat self-documenting, but that doesn't really work when you have short variable names. For example, you use ``gme`` to represent the Earth's effective curvature, but normally in Python you would call this variable something like ``earth_eff_curv``. I assume that this is because you adopted the variable names from the old C++ code.
>   - There are almost no inline comments, and since the variable names aren't obvious, it makes it very difficult to understand what's going on. I would suggest adding references. For example:
>      ```python
>      # Calculate attenuation
>      # Eqn. 6.3 in The ITS Irregular Terrain Model, the Algorithm - G. Hufford
>      ```
>   - The package is flat (no sub-directories) and each function is in its own separate module. To make it easier to navigate, you may want to consider grouping similar functions into modules and similar modules into sub-packages. Again, I assume that this is because you adopted the structure of some old MATLAB code.
>
> Overall, the package looks very useful. I think if you add a couple more comments to the ``scripts/`` and ``tests/`` directories, it will help new users to understand how the package works and ensure that the package is maintainable.

_Originally posted by @garrettj403 in https://github.com/openjournals/joss-reviews/issues/2266#issuecomment-647231325_
