### 2020-06-27 edwardoughton/itmlogic#45 @edwardoughton
@garrettj403 I've proposed a new folder structure in the next iteration of corrections, which resembles the same function/module grouping as in the box diagram in the paper. Hopefully this will make things clearer.
