### 2020-06-27 5a8699a
Add test doc strings
