+++ src/itmlogic/statistics/avar.py
@@ -1,0 +1,217 @@
+import math
+import numpy as np
+
+from itmlogic.curv import curv
+
+def avar(zzt, zzl, zzc, prop):
+    """
+    Finds the quantiles of attenuation using the output from lrprop.
+
+    When in area prediction mode we need a threefold quantile of attenuation corresponding
+    to time, locations and situations. For efficiency, avar is written as a function of the
+    standard normal deviates corresponding to the requested fractions.
+
+    Parameters
+    ----------
+    zzt : float
+        Standard normal deviate corresponding to user defined quantile.
+    zzl : float
+        Standard normal deviate corresponding to user defined quantile.
+    zzc : float
+        Standard normal deviate corresponding to user defined quantile.
+    prop : dict
+        Contains all input propagation parameters.
+
+    Returns
+    -------
+    avar1 : float
+        Additional attenuation from the median corresponding to the user defined quantiles
+        in time, location, and situation (Section 5 of "The ITS Irregular Terrain Model,
+        version 1.2.2: The Algorithm").
+    prop : dict
+        Contains all input and output propagation parameters.
+
+    """
+    third = 1 / 3
+
+    bv1 = [   -9.67,   -0.62,    1.26,   -9.21,   -0.62,   -0.39,    3.15]
+    bv2 = [    12.7,    9.19,    15.5,    9.05,    9.19,    2.86,   857.9]
+    xv1 = [ 144.9e3, 228.9e3, 262.6e3,  84.1e3, 228.9e3, 141.7e3, 2222.e3]
+    xv2 = [ 190.3e3, 205.2e3, 185.2e3, 101.1e3, 205.2e3, 315.9e3, 164.8e3]
+    xv3 = [ 133.8e3, 143.6e3,  99.8e3,  98.6e3, 143.6e3, 167.4e3, 116.3e3]
+    bsm1 = [    2.13,    2.66,    6.11,    1.98,    2.68,    6.86,    8.51]
+    bsm2 = [   159.5,    7.67,    6.65,   13.11,    7.16,   10.38,   169.8]
+    xsm1 = [ 762.2e3, 100.4e3, 138.2e3, 139.1e3,  93.7e3, 187.8e3, 609.8e3]
+    xsm2 = [ 123.6e3, 172.5e3, 242.2e3, 132.7e3, 186.8e3, 169.6e3, 119.9e3]
+    xsm3 = [  94.5e3, 136.4e3, 178.6e3, 193.5e3, 133.5e3, 108.9e3, 106.6e3]
+    bsp1 = [    2.11,    6.87,   10.08,    3.68,    4.75,    8.58,    8.43]
+    bsp2 = [   102.3,   15.53,    9.60,   159.3,    8.12,   13.97,    8.19]
+    xsp1 = [ 636.9e3, 138.7e3, 165.3e3, 464.4e3,  93.2e3, 216.0e3, 136.2e3]
+    xsp2 = [ 134.8e3, 143.7e3, 225.7e3,  93.1e3, 135.9e3, 152.0e3, 188.5e3]
+    xsp3 = [  95.6e3,  98.6e3, 129.7e3,  94.2e3, 113.4e3, 122.7e3, 122.9e3]
+    bsd1 = [   1.224,   0.801,   1.380,   1.000,   1.224,   1.518,   1.518]
+    bzd1 = [   1.282,   2.161,   1.282,     20.,   1.282,   1.282,   1.282]
+    bfm1 = [      1.,      1.,      1.,      1.,    0.92,      1.,      1.]
+    bfm2 = [      0.,      0.,      0.,      0.,    0.25,      0.,      0.]
[…]
