### 2020-06-27 240998c
Add test doc string
