### 2020-06-28 #2266 @edwardoughton
> 👋 @edwardoughton - This is just to make sure you've seen the comments and are working on responding to them...

Most of this is now complete @danielskatz, I'm just waiting for the authorship team to agree all the changes, and then I'll integrate the branch for @garrettj403 to review. I'm hoping this will be done by tomorrow.
