### 2020-06-28 6f9e604
Update inline comments for area + p2p modes
