+++ tests/test_ahd.py
@@ -7,6 +7,9 @@ def test_ahd():
     The Algorithm") used in the computation of tropospheric scatter attenuation,
     with the input D in meters.
 
+    The inputs and expected answer are based on an original test for Longley-Rice between
+    for Crystal Palace (South London) to Mursley, England (See Stark, 1967).
+
     """
     #td input in meters
     td = 5643.8
