+++ tests/test_alos.py
@@ -6,11 +6,11 @@ def test_alos(setup_prop_to_test_alos):
     Test the 'line-of-sight attenuation' at the distance d using a combination of plane
     earth fields and directed fields.
 
-    The imported setup parameters are imported from tests/conftest.py via the fixture
+    The setup parameters are imported from tests/conftest.py via the fixture
     setup_prop_to_test_alos.
 
-    The test is derived from the original test for Longley-Rice between for Crystal
-    Palace (South London) to Mursley, England.
+    The inputs and expected answer are based on an original test for Longley-Rice between
+    for Crystal Palace (South London) to Mursley, England (See Stark, 1967).
 
     """
     actual_answer = alos(
