+++ tests/test_area.py
@@ -81,86 +81,127 @@ def test_itmlogic_area():
     #(Average system elev above sea level)
     zsys = 0
 
+
+    #Rescale requested percentages into their corresponding normal distribution arguments
     ZT = qerfi([x / 100 for x in QT])[0]
     ZL = qerfi([x / 100 for x in QL])[0]
     ZC = qerfi([x / 100 for x in QC])
 
+    #The number of confidence intervals requested
     NC = len(QC)
 
+    #Don't allow negative distances
     if (D0 <= 0):
         D0 = DS1
+
+    #Set initial distance to 2 km if D0<=0
     if (D0 <= 0):
         D0= 2
+
+    #If final distance less than initial, only do one distance
     if (D1 <= D0) or (DS1 <= 0):
         ND = 1
         D1 = D0
+    #Otherwise compute the number of distance points in the loop
+    #and recompute the final distance using this grid
     else:
         DS = DS1
         ND = math.floor((D1 - D0) / DS + 1.75)
         D1 = D0 + (ND - 1) * DS
 
+    #Repeat these corrections for the "coarse" grid in range that follows the fine grid
     if (D2 <= D1 ) or ( DS2 <= 0):
-        NDC = 0
+        NDC = 0 #If input parameters are wrong, don't do a coarse grid
+    #Otherwise set up appropriate coarse grid
     else:
         NDC = ND
         ND = math.floor((D2 - D1) / DS2 + 0.75)
         D2 = D1 + ND * DS2
         ND = NDC + ND
 
+    #Standard Earth curvature parameter
     prop['gma'] = 157E-9
+    #Scale factor for converting Np/km to dB/km
     DB = 8.685890
+    #Scale factor to convert km to m
     AKM = 1000
 
+    #Initialize error flag to 0
     prop['kwx'] = 0
+    #Initialize the omega_n parameter
     prop['wn'] = prop['fmhz'] / 47.7
+    #Initialize the surface refractivity
     prop['ens']  = prop['ens0']
 
+    #Adjust surface refractivity parameter if zsys set by user
     if (zsys != 0):
         prop['ens'] = prop['ens'] * math.exp(-zsys / 9460)
 
+    #Implement refractive effects on Earth curvature
     prop['gme']  = prop['gma'] * (1 - 0.04665 * math.exp(prop['ens'] / 179.3))
 
-    # Initial value
+    #Initialize lvar parameter (this is used when updating AVAR with new input parameters)
     prop['lvar'] = 0
 
+    #Compute ground effective impedance zq parameter
     zq = complex(prop['eps'], 376.62 * prop['sgm'] / prop['wn'])
 
+    #Compute ground effective impedance z parameter (horizontal pol)
     prop['zgnd'] = math.sqrt(zq.real - 1)
 
+    #Compute ground effective impedance z parameter (vertical pol)
     if (prop['ipol'] != 0):
         prop['zgnd'] = prop['zgnd'] / zq
 
+    #Qlra initializes all the required parameters for area-prediction mode given
+    #siting type and other params already set in "prop"
     prop = qlra(KST, prop)
 
+    #Starting distances for loop over range
     D = D0
+    #First range step is that of the fine grid
     DT = DS
+
     FS = []
     DD = []
 
     output = []
+
     for JD in range(0, ND): #0-22
 
+        #Ensure that AVAR routines adjust only for distance (not quantiles which are set)
         prop['lvar'] = max(1, prop['lvar'])
 
+        #Compute baseline propagation loss at current range
         prop = lrprop(D * AKM, prop)
 
+        #Compute and store baseline loss
         FS.append(DB * np.log(2 * prop['wn'] * prop['dist']))
+
+        #Store distance at this increment
         DD.append(D)
 
+        #Loop over confidence intervals requested by user
[…]
