+++ tests/test_ascat.py
@@ -8,8 +8,8 @@ def test_ascat(setup_prop_to_test_ascat):
     The imported setup parameters are imported from tests/conftest.py via the fixture
     setup_prop_to_test_ascat.
 
-    The test is derived from the original test for Longley-Rice between for Crystal
-    Palace (South London) to Mursley, England.
+    The inputs and expected answer are based on an original test for Longley-Rice between
+    for Crystal Palace (South London) to Mursley, England (See Stark, 1967).
 
     """
     actual_prop = ascat(418934.4081874959, setup_prop_to_test_ascat)
