### 2020-06-29 6975e81
Enhance doc strings in module functions
