+++ src/itmlogic/scatter_attenuation/ascat.py
@@ -6,8 +6,8 @@ from itmlogic.ahd import ahd
 def ascat(d, prop):
     """
     Finds the scatter attenuation at the distance d using an approximation to the methods of
-    NBS TN101 with checks for inadmissable situations. A call with d = 0 set up initial
-    constants.
+    NBS TN101 with checks for inadmissable situations (Eqn 4.63 of "The ITS Irregular Terrain
+    Model, version 1.2.2: The Algorithm"). A call with d = 0 set up initial constants.
 
     Parameters
     ----------
