### 2020-06-29 fc8ed64
Enhance test doc strings in module functions
