+++ tests/test_ascat.py
@@ -3,7 +3,8 @@ from itmlogic.ascat import ascat
 
 def test_ascat(setup_prop_to_test_ascat):
     """
-    Test the scatter attenuation function ascat at the stated distance d.
+    Test the scatter attenuation function ascat at the stated distance d (Eqn 4.63 of "The
+    ITS Irregular Terrain Model, version 1.2.2: The Algorithm").
 
     The imported setup parameters are imported from tests/conftest.py via the fixture
     setup_prop_to_test_ascat.
@@ -14,4 +15,4 @@ def test_ascat(setup_prop_to_test_ascat):
     """
     actual_prop = ascat(418934.4081874959, setup_prop_to_test_ascat)
 
-    assert round(actual_prop['ascat1'],3) == 99.023
+    assert round(actual_prop['ascat1'], 3) == 99.023
