+++ tests/test_aknfe.py
@@ -3,7 +3,9 @@ from itmlogic.aknfe import aknfe
 
 def test_aknfe():
     """
-    Returns the attenuation due to a single knife edge - the Fresnal integral (in decibels).
+    Returns the attenuation due to a single knife edge - the Fresnel integral (in decibels,
+    Eqn 4.21 of "The ITS Irregular Terrain Model, version 1.2.2: The Algorithm" – see also
+    Eqn 6.1) evaluated for nu equal to the square root of the input argument.
 
     """
     assert round(aknfe(-1), 2) == 6.05
