### 2020-06-30 7a0ae8d
Update install / vis instructions
