### 2020-06-30 acce124
Update path to use results directory
