### 2020-06-30 edwardoughton/itmlogic#43 @edwardoughton
@garrettj403 Comments have now been added throughout both runner scripts (scripts/area.py and scripts/p2p.py), and doc strings have been improved throughout all modules and tests. We have also added visualized results to the readme, and provide all the code in the repo folder vis/vis.py for producing these plots. I also removed the typo you mentioned - thanks for catching that.
