### 2020-06-30 edwardoughton/itmlogic#44 @edwardoughton
@garrettj403 Doc strings have now been improved throughout all modules and tests. Where we have a direct reference for an equation in the references documentation, we have been explicit about it. Hopefully this makes things a bit clearer now.
