### 2020-06-30 edwardoughton/itmlogic#45 @edwardoughton
John (@garrettj403), we have now resolved:
- In-line comments throughout the runner scripts
- Doc string information in both modules and tests
- The flat structure (we now have a hierarchy which matches the diagram provided in the paper (for easy user reference)).

The only things we decided not to change currently were the variable names, as many users will be familiar with the names from the original code and we don't want to break from tradition in case it causes people not to use the repo. We decided this is something to consider in a future version, where a user may be able to specify if they want the traditional nomenclature or pythonic var names.
