### 2020-07-01 #2266 @garrettj403
Hi @edwardoughton, the changes look great! The additional comments will definitely help new users get started with your code.

@danielskatz, I'm ready to recommend publication. ``itmlogic`` was easy to install on my computer and all of the unit tests run smoothly without any errors. The package also has detailed documentation, which is hosted online. Overall, it looks like a very useful package with clear applications.

Great job!
