### 2020-07-01 #2266 @gorlapraveen
Hi @edwardoughton, the functionality and the use of algorithms is technically quite satisfying. However,  I would also suggest you to document where User inputs can be given. Such as where to change the operating frequencies for different modes.  This would be my final comment and we can proceed for publication.
