### 2020-07-02 609531f
Update area script with new dict
