+++ scripts/area_2tiles.py
@@ -1,0 +1,300 @@
+"""
+Testing a area prediction mode over two DEM tiles.
+
+Written by Tom Russell and Ed Oughton
+
+November 2019
+
+Original Matlab code implemented by Johnson and Yardim at OSU.
+
+"""
+import configparser
+import os
+import csv
+import math
+import numpy as np
+from functools import partial
+
+from itmlogic.misc.qerfi import qerfi
+from itmlogic.preparatory_subroutines.qlra import qlra
+from itmlogic.lrprop import lrprop
+from itmlogic.statistics.avar import avar
+from terrain_module import terrain_area
+
+# #set up file paths
+CONFIG = configparser.ConfigParser()
+CONFIG.read(os.path.join(os.path.dirname(__file__), 'script_config.ini'))
+BASE_PATH = CONFIG['file_locations']['base_path']
+
+DATA_PROCESSED = os.path.join(BASE_PATH, 'processed')
+RESULTS = os.path.join(BASE_PATH, '..', 'results')
+
+
+def itmlogic_area(main_user_defined_parameters):
+    """
+    Run itmlogic in area prediction mode.
+
+    Parameters
+    ----------
+    main_user_defined_parameters : dict
+        User defined parameters.
+
+    Returns
+    -------
+    output : list of dicts
+        Contains model output results.
+
+    """
+    prop = main_user_defined_parameters
+
+    #DEFINE ENVIRONMENTAL PARAMETERS
+    #Surface refractivity (N-units)
+    prop['ens0'] = 301
+
+    #Relative permittivity of ground
+    prop['eps'] = 15
+
+    #Conductivity (S/m) of ground
+    prop['sgm'] = 0.001
+
+    # Climate selection (1=equatorial, 2=continental subtropical, 3=maritime subtropical,
+    # 4=desert, 5=continental temperate, 6=maritime temperate overland,
+    # 7=maritime temperate, oversea (5 is the default)
+    prop['klimx'] = 5
+
+    #Mode of variability: Single Message=0, Accidental=1, Mobile=2, Broadcast=3
+    prop['mdvarx'] = 3
+
+    #DEFINE STATISTICAL PARAMETERS
+    #Percent of time requested for computation
+    QT = [50]
+
+    #Percent of locations requested for computation
+    QL = [50]
+
+    #Confidence levels of computation
+    QC = [50, 90, 10]
+
+    #Initial distance (km) for loop over range
+    D0 = 10
+
+    #Max distance 1 (km) for loop over range
+    D1 = 150
+
+    #Increment 1 (km) for loop over range
+    DS1 = 10
+
+    #Max distance 2 (km) for coarser loop over range (starts beyond D1)
+    D2 = 500
+
+    #Increment 2 (km) for coarser loop over range
+    DS2 = 50
+
+    #Siting criterion for antenna 1, 0=random, 1= careful, 2= very careful
+    KST = [2, 2]
+
+    #Refractivity scaling ens=ens0*exp(-zsys/9460.);
+    #(Average system elev above sea level)
+    zsys = 0
+
+    #Rescale requested percentages into their corresponding normal distribution arguments
+    ZT = qerfi([x / 100 for x in QT])[0]
+    ZL = qerfi([x / 100 for x in QL])[0]
+    ZC = qerfi([x / 100 for x in QC])
+
+    #The number of confidence intervals requested
+    NC = len(QC)
+
+    #Don't allow negative distances
+    if (D0 <= 0):
+        D0 = DS1
+
+    #Set initial distance to 2 km if D0<=0
+    if (D0 <= 0):
+        D0= 2
+
+    #If final distance less than initial, only do one distance
+    if (D1 <= D0) or (DS1 <= 0):
+        ND = 1
+        D1 = D0
+    #Otherwise compute the number of distance points in the loop
[…]
