+++ scripts/p2p.py
@@ -35,16 +35,17 @@ BASE_PATH = CONFIG['file_locations']['base_path']
 DATA_PROCESSED = os.path.join(BASE_PATH, 'processed')
 RESULTS = os.path.join(BASE_PATH, '..', 'results')
 
-def itmlogic_p2p(surface_profile_m, distance_km):
+
+def itmlogic_p2p(main_user_defined_parameters, surface_profile_m):
     """
     Run itmlogic in point to point (p2p) prediction mode.
 
     Parameters
     ----------
+    main_user_defined_parameters : dict
+        User defined parameters.
     surface_profile_m : list
         Contains surface profile measurements in meters.
-    distance_km : float
-        Distance between the transmitter and receiver.
 
     Returns
     -------
@@ -52,25 +53,14 @@ def itmlogic_p2p(surface_profile_m, distance_km):
         Contains model output results.
 
     """
-    prop = {}
+    prop = main_user_defined_parameters
 
+    #DEFINE ENVIRONMENTAL PARAMETERS
     # Terrain relative permittivity
     prop['eps'] = 15
 
     # Terrain conductivity (S/m)
     prop['sgm']   = 0.005
-
-    # Polarization selection (0=horizontal, 1=vertical)
-    prop['ipol'] = 0
-
-    # Operating frequency (MHz) % Second qkpfl.for test case
-    prop['fmhz'] = 573.3
-
-    # Antenna 1 height (m) # Antenna 2 height (m)
-    prop['hg'] = [143.9, 8.5]
-
-    # Operating frequency (MHz) % First qkpfl.for test case
-    prop['fmhz']  =  41.5
 
     # Climate selection (1=equatorial,
     # 2=continental subtropical, 3=maritime subtropical,
@@ -82,14 +72,12 @@ def itmlogic_p2p(surface_profile_m, distance_km):
     # Surface refractivity (N-units): also controls effective Earth radius
     prop['ens0']  =   314
 
+    #DEFINE STATISTICAL PARAMETERS
     # Confidence  levels for predictions
     qc = [50, 90, 10]
 
     # Reliability levels for predictions
     qr = [1, 10, 50, 90, 99]
-
-    # Length of profile
-    prop['d'] = distance_km
 
     # Number of points describing profile -1
     pfl = []
@@ -355,14 +343,30 @@ def straight_line_from_points(a, b):
 
 if __name__ == '__main__':
 
+    #setup data folder paths
     dem_folder = os.path.join(BASE_PATH)
     directory_shapes = os.path.join(DATA_PROCESSED, 'shapes')
 
+    #set coordinate reference systems
     old_crs = 'EPSG:4326'
-    new_crs = 'EPSG:3857'
+    # new_crs = 'EPSG:3857'
 
-    #original distance in km from Longley Rice docs
-    original_distance = 77.8
+    #DEFINE MAIN USER PARAMETERS
+    #define an empty dict for user defined parameters
+    main_user_defined_parameters = {}
+
+    #define radio operating frequency (MHz)
+    # main_user_defined_parameters['fmhz'] = 573.3
+    main_user_defined_parameters['fmhz']  =  41.5
+
+    #define distance between terminals in km (from Longley Rice docs)
+    main_user_defined_parameters['d'] = 77.8
+
+    #define antenna heights - Antenna 1 height (m) # Antenna 2 height (m)
+    main_user_defined_parameters['hg'] = [143.9, 8.5]
+
+    #polarization selection (0=horizontal, 1=vertical)
+    main_user_defined_parameters['ipol'] = 0
 
     #original surface profile from Longley Rice docs
     original_surface_profile_m = [
@@ -416,16 +420,13 @@ if __name__ == '__main__':
[…]
