+++ scripts/p2p_2tiles.py
@@ -1,0 +1,441 @@
+"""
+Testing a point to point path over two DEM tiles.
+
+Written by Tom Russell and Ed Oughton
+
+May 2020
+
+"""
+import configparser
+import os
+import csv
+import math
+import numpy as np
+from functools import partial
+from collections import OrderedDict
+
+import fiona
+from fiona.crs import from_epsg
+from pyproj import Transformer
+from shapely.geometry import LineString, mapping
+from shapely.ops import transform
+
+from itmlogic.misc.qerfi import qerfi
+from itmlogic.preparatory_subroutines.qlrpfl import qlrpfl
+from itmlogic.statistics.avar import avar
+from terrain_module import terrain_p2p
+
+# #set up file paths
+CONFIG = configparser.ConfigParser()
+CONFIG.read(os.path.join(os.path.dirname(__file__), 'script_config.ini'))
+BASE_PATH = CONFIG['file_locations']['base_path']
+
+DATA_PROCESSED = os.path.join(BASE_PATH, 'processed')
+RESULTS = os.path.join(BASE_PATH, '..', 'results')
+
+
+def itmlogic_p2p(main_user_defined_parameters, surface_profile_m):
+    """
+    Run itmlogic in point to point (p2p) prediction mode.
+
+    Parameters
+    ----------
+    main_user_defined_parameters : dict
+        User defined parameters.
+    surface_profile_m : list
+        Contains surface profile measurements in meters.
+
+    Returns
+    -------
+    output : list of dicts
+        Contains model output results.
+
+    """
+    prop = main_user_defined_parameters
+
+    #DEFINE ENVIRONMENTAL PARAMETERS
+    # Terrain relative permittivity
+    prop['eps'] = 15
+
+    # Terrain conductivity (S/m)
+    prop['sgm']   = 0.005
+
+    # Climate selection (1=equatorial,
+    # 2=continental subtropical, 3=maritime subtropical,
+    # 4=desert, 5=continental temperate,
+    # 6=maritime temperate overland,
+    # 7=maritime temperate, oversea (5 is the default)
+    prop['klim']  =   5
+
+    # Surface refractivity (N-units): also controls effective Earth radius
+    prop['ens0']  =   314
+
+    #DEFINE STATISTICAL PARAMETERS
+    # Confidence  levels for predictions
+    qc = [50, 90, 10]
+
+    # Reliability levels for predictions
+    qr = [1, 10, 50, 90, 99]
+
+    # Number of points describing profile -1
+    pfl = []
+    pfl.append(len(surface_profile_m) - 1)
+    pfl.append(0)
+
+    for profile in surface_profile_m:
+        pfl.append(profile)
+
+    # Refractivity scaling ens=ens0*exp(-zsys/9460.)
+    # (Average system elev above sea level)
+    zsys = 0
+
+    # Note also defaults to a continental temperate climate
+
+    # Setup some intermediate quantities
+    # Initial values for AVAR control parameter: LVAR=0 for quantile change,
+    # 1 for dist change, 2 for HE change, 3 for WN change, 4 for MDVAR change,
+    # 5 for KLIM change
+    prop['lvar'] = 5
+
+    # Inverse Earth radius
+    prop['gma']  = 157E-9
+
+    # Conversion factor to db
+    db = 8.685890
+
+    #Number of confidence intervals requested
+    nc = len(qc)
+
+    #Number of reliability intervals requested
+    nr = len(qr)
+
+    #Length of profile in km
+    dkm = prop['d']
+
+    #Profile range step, select option here to define range step from profile
+    #length and # of points
+    xkm = 0
+
+    #If DKM set <=0, find DKM by mutiplying the profile step by number of
+    #points (not used here)
+    if dkm <= 0:
+        dkm = xkm * pfl[0]
[…]
