### 2020-07-02 #2266 @edwardoughton
@gorlapraveen I've now adapted the example running scripts and the documentation to reflect the main user defined parameters. This includes having a dict which contains the main parameters which the user can concisely state before calling the ``itmfunction``. Hopefully this makes things clearer.

I also separated out the example scripts which use either 1 DEM tile or 2 DEM tiles, to overcome the issue of working across large areas which some users will need to deal with. The documentation also now reflects these changes.

Thanks for the review feedback, Ed
