### 2020-07-05 #2266 @danielskatz
@edwardoughton  - at this point, please:
- [ ] Archive the reviewed software in Zenodo or a similar service (e.g. figshare, an institutional repository)
- [ ] Check the archival deposit (e.g., in Zenodo) has the correct metadata, this includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it); you may also add the authors' ORCID.

I can then move forward with accepting the submission, which will include me proofreading the paper and the references.
