### 2020-07-06 #2266 @edwardoughton
> @edwardoughton - at this point, please:
>
> * [x]  Make a tagged release of your software, and list the version tag of the archived version here.
> * [x]  Archive the reviewed software in Zenodo or a similar service (e.g. figshare, an institutional repository)
> * [x]  Check the archival deposit (e.g., in Zenodo) has the correct metadata, this includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it); you may also add the authors' ORCID.
> * [ ]  Please list the DOI of the archived version here.
>
> I can then move forward with accepting the submission, which will include me proofreading the paper and the references.

Thanks @danielskatz, I should now have completed those tasks. Please let me know if I missed anything.

DOI: 10.5281/zenodo.3931350

Zonodo link: https://zenodo.org/record/3931350#.XwLumihKh9M
