### 2020-07-06 #2266 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
